Seen_Pipeline/pvnet_lnd/requirements.txt
yacs
numpy==1.22.0
#torchvision==0.2.1
opencv-python
tqdm==4.28.1
pycocotools==2.0.0
matplotlib==2.2.2
plyfile==0.6
scikit-image==0.14.2
scikit-learn
PyOpenGL==3.1.1a1
ipdb
cyglfw3==3.1.0.2
pyassimp==3.3
progressbar
open3d-python
tensorboardX==1.2
cffi
